스킬 업그레이드 기본 로직 작성함. 이제 ui를 닫을 때 초기화하는거랑 그래픽적으로 추가해야한다.

+++ 02. Script/Character/CharacterData.cs
@@ -21,11 +21,12 @@ public class CharacterData:MonoBehaviour
     public int gridIndex;
     public Dictionary<EffectType, float> PermEffects { get; private set; }//º¸·ù
     public SkillAsItem[] skillAsIItems;
-    public float[] exp;
+    public int[] exp;
     public WeaponClass weapon;
     public CharacterHierarchy characterHierarchy;
     public CharacterInBattle characterAtBattle;
-    internal void InitCharacterData(string _docId, string _jobId, float _maxHp, float _hp, float _ability, float _resist, float _speed, int _gridIndex, SkillAsItem[] _skillasItems, float[] _exp, WeaponClass _weapon)
+    internal void InitCharacterData(string _docId, string _jobId, float _maxHp, float _hp, float _ability,
+        float _resist, float _speed, int _gridIndex, SkillAsItem[] _skillasItems, int[] _exp, WeaponClass _weapon)
     {
         docId = _docId;
         jobClass = LoadManager.loadManager.jobsDict[_jobId];
